tests/fh/test_fh_cart_verify_cart_summary_section.py
"""
Cart Test - Migrated from Selenium

Test ID: T346
Original: Cart_SummarySection_TS_T346.java
"""
import pytest
import allure
from playwright.sync_api import Page

pytestmark = [pytest.mark.fourhands, pytest.mark.cart]


@allure.feature("Cart")
@allure.story("Verify Cart Summary Section")
@allure.title("T346: Verify Cart Summary Section")
@pytest.mark.smoke
def test_verify_cart_summary_section(fh_authenticated_page: Page):
    """
    Test: Verify Cart Summary Section
    
    Test ID: T346
    
    Steps:
    
    
    Migrated from: Cart_SummarySection_TS_T346.java
    """
    page = fh_authenticated_page
    
    # TODO: Implement test steps
    # This test was auto-generated and needs manual review
    
    pass  # TODO: Add test steps
    
    # TODO: Add assertions
    assert True, "Test needs implementation"

