requirements.txt
playwright>=1.40.0
pytest>=7.4.0
pytest-playwright>=0.4.3
pytest-xdist>=3.3.1
pytest-rerunfailures>=12.0
python-dotenv>=1.0.0
allure-pytest>=2.13.2
requests>=2.31.0
browserstack-sdk>=1.6.0
